+++ src/test/java/org/folio/search/service/consortia/FeatureConfigServiceDecoratorTest.java
@@ -27,7 +27,7 @@ class FeatureConfigServiceDecoratorTest extends DecoratorBaseTest {
 
   @Test
   void isEnabled() {
-    var feature = TenantConfiguredFeature.CONSORTIUM;
+    var feature = TenantConfiguredFeature.SEARCH_ALL_FIELDS;
     when(service.isEnabled(feature)).thenReturn(true);
     mockExecutor(consortiaTenantExecutor);
 
@@ -66,7 +66,7 @@ class FeatureConfigServiceDecoratorTest extends DecoratorBaseTest {
 
   @Test
   void update() {
-    var feature = TenantConfiguredFeature.CONSORTIUM;
+    var feature = TenantConfiguredFeature.SEARCH_ALL_FIELDS;
     var expected = new FeatureConfig();
     when(service.update(feature, expected)).thenReturn(expected);
     mockExecutor(consortiaTenantExecutor);
@@ -80,7 +80,7 @@ class FeatureConfigServiceDecoratorTest extends DecoratorBaseTest {
 
   @Test
   void delete() {
-    var feature = TenantConfiguredFeature.CONSORTIUM;
+    var feature = TenantConfiguredFeature.SEARCH_ALL_FIELDS;
     mockExecutorRun(consortiaTenantExecutor);
 
     decorator.delete(feature);